Исходные код задачи 2. Чтобы не было неопределённости со знаковым битом во время сдвига разрядов числа влево, тип данных int заменён на unsigned int.

+++ dz_5_2.cpp
@@ -6,8 +6,8 @@ int main(int argc, char * argv [])
     #define ELEMENTS_COUNT 10
     
     int array[ELEMENTS_COUNT]; // Для генерации значений похожих на случайные, не инициализируем массив сразу, а сначала оставляем неопределённые значения
-    int minimum = ((int)-1) & ~((int)1 << ((sizeof(int) <<  3) - 1)); // Инициализируем максимальным возможным значением знакового типа данных "int"
-    int maximum = ((int)1 << ((sizeof(int) <<  3) - 1)); // Инициализируем минимальным возможным значением знакового типа данных "int"
+    int minimum = ((int)-1) & ~((unsigned int)1 << ((sizeof(int) <<  3) - 1)); // Инициализируем максимальным возможным значением знакового типа данных "int"
+    int maximum = (int)((unsigned int)1 << ((sizeof(int) <<  3) - 1)); // Инициализируем минимальным возможным значением знакового типа данных "int"
 
     for(int index = 0; index < ELEMENTS_COUNT; index++)
     {
